refactor(site): extract CopyableBlock and an inline-code renderer
Pulls the copy-enabled code block out of usage-tabs into its own
component files so it can be reused by the upcoming HighlighterTabs.
Also adds renderInlineCode, a tiny helper that turns backtick-wrapped
runs in a tab's content string into <code> elements, and wires it into
UsageTabs.
