+++ apps/site/components/copyable-block.types.ts
@@ -1,0 +1,4 @@
+export type CopyableBlockProps = {
+	html: string;
+	label?: string;
+};
